generating CCG parses

+++ src/scripts/kinect_vectors/merge_data.sh
@@ -1,23 +1,28 @@
 #!/bin/bash -e
 
-DIRS="data/cobot/set/kinect/0*"
+# DIRS="data/cobot/set/kinect/0*"
+DIRS="data/cobot/set/language_geography/small_domains/*"
 
 for f in $DIRS
 do
-    IN="$f/training.annotated.txt"
-    PARSED="$f/training.annotated.txt.parsed"
-    CCG="$f/training.annotated.txt.ccg"
-    MERGED="$f/training.annotated.txt.merged"
+    IN="$f/training.txt"
+    PARSED="$f/training.txt.parsed"
+    CCG="$f/training.txt.ccg"
+    MERGED="$f/training.txt.merged"
 
     TEMP="$f/training.temp.txt"
     TEMP2="$f/training.temp2.txt"
 
-    grep -v '^\*' $IN | sed 's/\([^)]\)$/\1;/g' > $TEMP
-    cut -f2 $CCG > $TEMP2
+    if [ -e $IN ]
+    then
+	echo $IN
 
-    paste -d ';' $TEMP $PARSED $TEMP2 > $MERGED
+	grep -v '^\*' $IN | grep -v '^$' | sed 's/\([^)]\)$/\1;/g' > $TEMP
+	cut -f2 $CCG > $TEMP2
 
-    rm $TEMP
-    rm $TEMP2
+	paste -d ';' $TEMP $PARSED $TEMP2 > $MERGED
 
+	rm $TEMP
+	rm $TEMP2
+    fi
 done
